refactor(datatrak): move DraftExistsModal to features/Survey/hooks
Move DraftExistsModal and useDraftExistsModal out of SurveySelectPage
into the shared features/Survey/hooks folder since they are now used by
both SurveySelectPage and LandingPage.
