+++ packages/datatrak-web/src/features/Survey/hooks/useDraftExistsModal.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useSurveyResponseDrafts } from '../../api/queries/useSurveyResponseDrafts';
+import { useSurveyResponseDrafts } from '../../../api/queries/useSurveyResponseDrafts';
 
 /**
  * Core hook for draft-exists modal logic.
