+++ packages/datatrak-web/src/views/LandingPage/RecentSurveysSection.tsx
@@ -9,8 +9,8 @@ import { useCurrentUserRecentSurveys } from '../../api';
 import { InlineScrollView, SurveyIcon, Tile, TileSkeleton } from '../../components';
 import { TileProps } from '../../components/Tile';
 import { useIsMobile } from '../../utils';
-import { DraftExistsModal } from '../SurveySelectPage/DraftExistsModal';
-import { useDraftExistsModal } from '../SurveySelectPage/useDraftExistsModal';
+import { DraftExistsModal } from '../../features/Survey/hooks/DraftExistsModal';
+import { useDraftExistsModal } from '../../features/Survey/hooks/useDraftExistsModal';
 import { SectionHeading } from './SectionHeading';
 
 const RecentSurveys = styled.section`
