+++ packages/datatrak-web/src/views/SurveySelectPage/SurveySelectPage.tsx
@@ -10,7 +10,7 @@ import { useUserCountries } from '../../features/CountrySelector/useUserCountrie
 import { Survey } from '../../types';
 import { useIsMobile } from '../../utils';
 import { DesktopTemplate } from './DesktopTemplate';
-import { DraftExistsModal } from './DraftExistsModal';
+import { DraftExistsModal } from '../../features/Survey/hooks/DraftExistsModal';
 import { MobileTemplate } from './MobileTemplate';
 import { useSurveySelectionWithDrafts } from './useSurveySelectionWithDrafts';
 import { useSyncProjectFromUrl } from './useSyncProjectFromUrl';
